feature(typescript): add a minimum working version with TypeScript
