+++ .eslintrc.json
@@ -1,6 +1,7 @@
 {
-  "extends": ["react-app", "react-jam3"],
-  "plugins": ["jam3"],
+  "extends": ["react-app", "react-jam3", "plugin:import/typescript"],
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["@typescript-eslint", "jam3"],
   "settings": {
     "react": {
       "version": "detect"
